requirements.txt
allure-pytest==2.9.45
Faker==13.0.0
pytest==7.0.1
pytest-xdist==2.5.0
selenium==4.1.0
